refactoring to non duplicate-code, adding of docs, improve pep8
