+++ NetCatKS/Config/api/implementers/configuration/mixin/__init__.py
@@ -15,11 +15,91 @@ from NetCatKS.DProtocol.api.public.actions import BaseProtocolActions
 __author__ = 'dimd'
 
 
-class MixinConfiguration(object):
+class MixinSharedConfig(BaseProtocolActions):
     """
-    A mixin configuration for a TCP and WEB servers
+    Represents a shared mixin configuration between all sections
     """
-    pass
+    def __init__(self):
+
+        """
+        We providing a default service name, port and hostname
+
+        :return: void
+        """
+        self.__service_name = 'DefaultServiceName'
+        self.__port = 8080
+        self.__hostname = '127.0.0.1'
+
+    @property
+    def service_name(self):
+
+        """
+        A Service Name getter for this TCP service, it used inside out Twisted TCP service.Service
+
+        :return: strs
+
+        """
+
+        return self.__service_name
+
+    @service_name.setter
+    def service_name(self, name):
+
+        """
+        A setter for a service name
+        :param name:
+        :type name: str
+
+        :return: void
+        """
+
+        self.__service_name = name
+
+    @property
+    def port(self):
+
+        """
+        Getter for a Service port
+        :return: int
+        """
+
+        return self.__port
+
+    @port.setter
+    def port(self, port):
+
+        """
+        A Service port setter we accept only integers for a port number
+
+        :param port:
+        :type port: int
+
+        :return: void
+
+        """
+
+        self.__port = port
+
+    @property
+    def hostname(self):
+        """
+        A WS hostname getter
+        :return:
+        """
+        return self.__hostname
+
+    @hostname.setter
+    def hostname(self, hostname):
+
+        """
+        A WS hostname setter
+
+        :param hostname:
+        :type hostname: str
+
+        :return: void
+        """
+        self.__hostname = hostname
 
 
 class RegisterAsFactory(object):
