+++ NetCatKS/Config/api/implementers/configuration/web/__init__.py
@@ -4,15 +4,16 @@ This module is written in a DynamicProtocol style and caring for our WEB Server 
 
 from zope.interface import implementer
 
-from NetCatKS.DProtocol.api.public.actions import BaseProtocolActions
 from NetCatKS.Config.api.interfaces import IWeb
+
+from NetCatKS.Config.api.implementers.configuration.mixin import MixinSharedConfig
 from NetCatKS.Config.api.implementers.configuration.mixin import RegisterAsFactory
 
 __author__ = 'dimd'
 
 
 @implementer(IWeb)
-class WEB(BaseProtocolActions):
+class WEB(MixinSharedConfig):
 
     """
     A Class representing our base WEB configuration
@@ -28,34 +29,12 @@ class WEB(BaseProtocolActions):
 
         :return: void
         """
+        super(WEB, self).__init__()
 
-        self.__service_name = 'Default WEB Server'
+        self.service_name = 'Default WEB Server'
+        self.port = 8000
         self.__http_methods = ['GET']
-        self.__port = 8000
         self.__www_root = ''
-
-    @property
-    def service_name(self):
-
-        """
-        Will return our web service name
-
-        :return: str
-        """
-
-        return self.__service_name
-
-    @service_name.setter
-    def service_name(self, sname):
-
-        """
-        A setter for our WEB service name
-        :param sname:
-        :type sname: str
-
-        :return: void
-        """
-        self.__service_name = sname
 
     @property
     def http_methods(self):
@@ -81,29 +60,6 @@ class WEB(BaseProtocolActions):
         """
 
         self.__http_methods = methods
-
-    @property
-    def port(self):
-
-        """
-        A WEB port gettger
-
-        :return: int
-
-        """
-        return self.__port
-
-    @port.setter
-    def port(self, port):
-
-        """
-
-        :param port:
-        :type port: int
-
-        :return: void
-        """
-        self.__port = port
 
     @property
     def www_root(self):
